fix: 모임 정보 조회 response 수정
- Meeting: 공유 가능 여부 필드 추가, 모임 개설 시점에 참가비 객체(Payment) 저장
- Payment: 참가비 타입을 Integer로 수정, builder() 수정
- MeetingService: 모임 개설 시 참가비 객체 생성 및 저장
- MeetingInfoDTO: 모임 정보 조회 API response 타입 수정
- PaymentDTO: 참가비 타입을 Integer로 수정

+++ back-end/src/main/java/kr/co/ssalon/web/dto/PaymentDTO.java
@@ -20,7 +20,7 @@ public class PaymentDTO {
     @NotBlank
     private Long meetingId;
     @NotBlank
-    private BigDecimal amount;
+    private Integer amount;
     @NotBlank
     private String purpose;
